Compile user themes at load time
User theme files are now cached in the local data storage path. When a user changes their theme it will be compiled once by stylus and then cached. If the cache file is missing at load it will compile it. There is a bit of hackery to modify the `url()` paths to make sure they are still relative to the app's source.